+++ TINetworking/Sources/Parameters/Encoding/QueryStringParameterEncoding.swift
@@ -25,10 +25,10 @@ open class QueryStringParameterEncoding: BaseUrlParameterEncoding, ParameterEnco
     // MARK: - ParameterEncoding
 
     open func encode(parameters: [String: Parameter<LocationQuery>]) -> [String: Any] {
-        let includedKeys = Set(super.encode(parameters: parameters).map { $0.key })
+        let nonEmptyValueKeys = Set(super.encode(parameters: parameters).map { $0.key })
 
         return parameters
-            .filter { includedKeys.contains($0.key) }
-            .mapValues { $0.value }
+            .filter { nonEmptyValueKeys.contains($0.key) }
+            .compactMapValues { $0.value }
     }
 }