+++ project-scripts/push_to_podspecs.sh
@@ -5,10 +5,15 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 cd "$DIR"
 
-# Push changes
-find ../ -name '*.podspec' \
-	-not -path "../Carthage/*" \
-	-not -path "../*/Carthage/*" \
-	-not -path "../Pods/*" \
-	-not -path "../*/Pods/*" \
-	| xargs -I% pod repo push git@github.com:TouchInstinct/Podspecs % --allow-warnings
+ORDERED_PODSPECS="../TISwiftUtils/TISwiftUtils.podspec
+../TIFoundationUtils/TIFoundationUtils.podspec
+../TIKeychainUtils/TIKeychainUtils.podspec
+../TIUIKitCore/TIUIKitCore.podspec
+../TIUIElements/TIUIElements.podspec
+../TITableKitUtils/TITableKitUtils.podspec
+../TINetworking/TINetworking.podspec
+../TIMoyaNetworking/TIMoyaNetworking.podspec"
+
+for podspec_path in ${ORDERED_PODSPECS}; do
+    pod repo push git@github.com:TouchInstinct/Podspecs ${podspec_path} --allow-warnings
+done